pyproject.toml
[project]
name = "momo-website"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pyyaml>=6.0.3",
]
